+++ packages/plan/src/Map/__tests__/MapSpoilPlan.spec.ts
@@ -1,4 +1,4 @@
-import { MapSpoilPlan } from '../MapSpoilPlan.js';
+import { MapSpoilPlan } from '../MapSpoilPlan';
 
 describe('MapSpoilPlan', () => {
   describe('of', () => {