+++ packages/plan/src/Recovery/RecoveryPlan.ts
@@ -1,5 +1,5 @@
-import { Noun, Whatever } from '@jamashita/anden-type';
+import { Noun } from '@jamashita/anden-type';
 
 export interface RecoveryPlan<R, N extends string = string> extends Noun<N> {
-  onRecover(value: R): Whatever;
+  onRecover(value: R): unknown;
 }